Trait wiggle::GuestErrorType

pub trait GuestErrorType {
    fn success() -> Self;
}

A trait for types which are used to report errors. Each type used in the first result position of an interface function is used, by convention, to indicate whether the function was successful and subsequent results are valid, or whether an error occured. This trait allows wiggle to return the correct value when the interface function’s idiomatic Rust method returns Ok().

Required methods

fn success() -> Self