Struct wiggle::wasmtime_crate::component::ExportInstance

pub struct ExportInstance<'a, 'store> { /* private fields */ }

Description of the exports of a single instance.

This structure is created from Exports via the Exports::root or Exports::instance methods. This type provides access to the first layer of exports within an instance. The ExportInstance::instance method can be used to provide nested access to sub-instances.

Implementations

impl<'a, 'store> ExportInstance<'a, 'store>

pub fn func(&mut self, name: &str) -> Option<Func>

Same as Instance::get_func

pub fn typed_func<Params, Results>( &mut self, name: &str ) -> Result<TypedFunc<Params, Results>, Error>where Params: ComponentNamedList + Lower, Results: ComponentNamedList + Lift,

Same as Instance::get_typed_func

pub fn module(&mut self, name: &str) -> Option<&'a Module>

Same as Instance::get_module

pub fn resource(&mut self, name: &str) -> Option<ResourceType>

Same as Instance::get_resource

pub fn modules(&self) -> impl Iterator<Item = (&'a str, &'a Module)>

Returns an iterator of all of the exported modules that this instance contains.

pub fn instance(&mut self, name: &str) -> Option<ExportInstance<'a, '_>>

Looks up the exported instance with the name specified and returns a description of its exports.

pub fn into_instance(self, name: &str) -> Option<ExportInstance<'a, 'store>>

Same as ExportInstance::instance but consumes self to yield a return value with the same lifetimes.