Struct wiggle::wasmtime_crate::component::Func

pub struct Func(_);

A WebAssembly component function.

Implementations

impl Func

pub fn typed<Params, Return>( &self, store: impl AsContext ) -> Result<TypedFunc<Params, Return>, Error>where Params: ComponentNamedList + Lower, Return: ComponentNamedList + Lift,

Attempt to cast this Func to a statically typed TypedFunc with the provided Params and Return.

This function will perform a type-check at runtime that the Func takes Params as parameters and returns Return. If the type-check passes then a TypedFunc will be returned which can be used to invoke the function in an efficient, statically-typed, and ergonomic manner.

The Params type parameter here is a tuple of the parameters to the function. A function which takes no arguments should use (), a function with one argument should use (T,), etc. Note that all Params must also implement the Lower trait since they’re going tin to wasm.

The Return type parameter is the return value of this function. A return value of () means that there’s no return (similar to a Rust unit return) and otherwise a type T can be specified. Note that the Return must also implement the Lift trait since it’s coming from wasm.

Types specified here must implement the ComponentType trait. This trait is implemented for built-in types to Rust such as integer primitives, floats, Option<T>, Result<T, E>, strings, Vec<T>, and more. As parameters you’ll be passing native Rust types.

Errors

If the function does not actually take Params as its parameters or return Return then an error will be returned.

Panics

This function will panic if self is not owned by the store specified.

Examples

Calling a function which takes no parameters and has no return value:

let typed = func.typed::<(), ()>(&store)?;
typed.call(store, ())?;

Calling a function which takes one string parameter and returns a string:

let typed = func.typed::<(&str,), (String,)>(&store)?;
let ret = typed.call(&mut store, ("Hello, ",))?.0;
println!("returned string was: {}", ret);

Calling a function which takes multiple parameters and returns a boolean:

let typed = func.typed::<(u32, Option<&str>, &[u8]), (bool,)>(&store)?;
let ok: bool = typed.call(&mut store, (1, Some("hello"), b"bytes!"))?.0;
println!("return value was: {ok}");

pub fn params(&self, store: impl AsContext) -> Box<[Type], Global>

Get the parameter types for this function.

pub fn results(&self, store: impl AsContext) -> Box<[Type], Global>

Get the result types for this function.

pub fn call( &self, store: impl AsContextMut, params: &[Val], results: &mut [Val] ) -> Result<(), Error>

Invokes this function with the params given and returns the result.

The params here must match the type signature of this Func, or this will return an error. If a trap occurs while executing this function, then an error will also be returned.

pub async fn call_async<T>( &self, store: impl AsContextMut<Data = T>, params: &[Val], results: &mut [Val] ) -> impl Future<Output = Result<(), Error>>where T: Send,

Exactly like Self::call except for use on async stores.

Panics

Panics if this is called on a function in a synchronous store. This only works with functions defined within an asynchronous store. Also panics if store does not own this function.

pub fn post_return(&self, store: impl AsContextMut) -> Result<(), Error>

Invokes the post-return canonical ABI option, if specified, after a Func::call has finished.

For some more information on when to use this function see the documentation for post-return in the Func::call method. Otherwise though this function is a required method call after a Func::call completes successfully. After the embedder has finished processing the return value then this function must be invoked.

Errors

This function will return an error in the case of a WebAssembly trap happening during the execution of the post-return function, if specified.

Panics

This function will panic if it’s not called under the correct conditions. This can only be called after a previous invocation of Func::call completes successfully, and this function can only be called for the same Func that was call’d.

If this function is called when Func::call was not previously called, then it will panic. If a different Func for the same component instance was invoked then this function will also panic because the post-return needs to happen for the other function.

Panics if this is called on a function in an asynchronous store. This only works with functions defined within a synchronous store.

pub async fn post_return_async<T>( &self, store: impl AsContextMut<Data = T> ) -> impl Future<Output = Result<(), Error>>where T: Send,

Exactly like Self::post_return except for use on async stores.

Panics

Panics if this is called on a function in a synchronous store. This only works with functions defined within an asynchronous store.