Enum wiggle::wasmtime_crate::HeapType

pub enum HeapType {
    Extern,
    Func,
    Concrete(FuncType),
    NoFunc,
}

The heap types that can Wasm can have references to.

Subtyping and Equality

HeapType does not implement Eq, because heap types have a subtyping relationship, and so 99.99% of the time you actually want to check whether one type matches (i.e. is a subtype of) another type. You can use the HeapType::matches method to perform these types of checks. If, however, you are in that 0.01% scenario where you need to check precise equality between types, you can use the HeapType::eq method.

Variants

Extern

The extern heap type represents external host data.

Func

The func heap type represents a reference to any kind of function.

This is the top type for the function references type hierarchy, and is therefore a supertype of every function reference.

Concrete(FuncType)

The concrete heap type represents a reference to a function of a specific, concrete type.

This is a subtype of func and a supertype of nofunc.

NoFunc

The nofunc heap type represents the null function reference.

This is the bottom type for the function references type hierarchy, and therefore nofunc is a subtype of all function reference types.

Implementations

impl HeapType

pub fn is_extern(&self) -> bool

Is this the abstract extern heap type?

pub fn is_func(&self) -> bool

Is this the abstract func heap type?

pub fn is_no_func(&self) -> bool

Is this the abstract nofunc heap type?

pub fn is_abstract(&self) -> bool

Is this an abstract type?

Types that are not abstract are concrete, user-defined types.

pub fn is_concrete(&self) -> bool

Is this a concrete, user-defined heap type?

Types that are not concrete, user-defined types are abstract types.

pub fn as_concrete(&self) -> Option<&FuncType>

Get the underlying concrete, user-defined type, if any.

Returns None for abstract types.

pub fn unwrap_concrete(&self) -> &FuncType

Get the underlying concrete, user-defined type, panicking if this heap type is not concrete.

pub fn top(&self, engine: &Engine) -> HeapType

Get the top type of this heap type’s type hierarchy.

The returned heap type is a supertype of all types in this heap type’s type hierarchy.

pub fn matches(&self, other: &HeapType) -> bool

Does this heap type match the other heap type?

That is, is this heap type a subtype of the other?

Panics

Panics if either type is associated with a different engine from the other.

pub fn eq(a: &HeapType, b: &HeapType) -> bool

Is heap type a precisely equal to heap type b?

Returns false even if a is a subtype of b or vice versa, if they are not exactly the same heap type.

Panics

Panics if either type is associated with a different engine from the other.