Enum wiggle::wasmtime_crate::Ref

pub enum Ref {
    Func(Option<Func>),
    Extern(Option<Rooted<ExternRef>>),
}

A reference.

References come in three broad flavors:

Function references. These are references to a function that can be invoked.

External references. These are references to data that is external and opaque to the Wasm guest, provided by the host.

Internal references. These are references to allocations inside the Wasm’s heap, such as structs and arrays. These are part of the GC proposal, and not yet implemented in Wasmtime.

At the Wasm level, there are nullable and non-nullable variants of each type of reference. Both variants are represented with Ref at the Wasmtime API level. For example, values of both (ref extern) and (ref null extern) types will be represented as Ref::Extern(Option<ExternRef>) in the Wasmtime API. Nullable references are represented as Option<Ref> where null references are represented as None. Wasm can construct null references via the ref.null <heap-type> instruction.

References are non-forgable: Wasm cannot create invalid references, for example, by claiming that the integer 0xbad1bad2 is actually a reference.

Variants

Func(Option<Func>)

A first-class reference to a WebAssembly function.

The host, or the Wasm guest, can invoke this function.

The host can create function references via Func::new or Func::wrap.

The Wasm guest can create non-null function references via the ref.func instruction, or null references via the ref.null func instruction.

Extern(Option<Rooted<ExternRef>>)

A reference to an value outside of the Wasm heap.

These references are opaque to the Wasm itself. Wasm can’t create non-null external references, nor do anything with them accept pass them around as function arguments and returns and place them into globals and tables.

Wasm can create null external references via the ref.null extern instruction.

Implementations

impl Ref

pub fn is_null(&self) -> bool

Is this a null reference?

pub fn is_non_null(&self) -> bool

Is this a non-null reference?

pub fn is_extern(&self) -> bool

Is this an extern reference?

pub fn as_extern(&self) -> Option<Option<&Rooted<ExternRef>>>

Get the underlying extern reference, if any.

Returns None if this Ref is not an extern reference, eg it is a func reference.

Returns Some(None) if this Ref is a null extern reference.

Returns Some(Some(_)) if this Ref is a non-null extern reference.

pub fn unwrap_extern(&self) -> Option<&Rooted<ExternRef>>

Get the underlying extern reference, panicking if this is a different kind of reference.

Returns None if this Ref is a null extern reference.

Returns Some(_) if this Ref is a non-null extern reference.

pub fn is_func(&self) -> bool

Is this a func reference?

pub fn as_func(&self) -> Option<Option<&Func>>

Get the underlying func reference, if any.

Returns None if this Ref is not an func reference, eg it is an extern reference.

Returns Some(None) if this Ref is a null func reference.

Returns Some(Some(_)) if this Ref is a non-null func reference.

pub fn unwrap_func(&self) -> Option<&Func>

Get the underlying func reference, panicking if this is a different kind of reference.

Returns None if this Ref is a null func reference.

Returns Some(_) if this Ref is a non-null func reference.

pub fn ty(&self, store: impl AsContext) -> RefType

Get the type of this reference.

Panics

Panics if this reference is associated with a different store.

pub fn matches_ty(&self, store: impl AsContext, ty: &RefType) -> bool

Does this reference value match the given type?

Panics

Panics if this reference is not associated with the given store.