Struct wiggle::wasmtime_crate::Instance

#[repr(transparent)]
pub struct Instance(_);

An instantiated WebAssembly module.

This type represents the instantiation of a Module. Once instantiated you can access the exports which are of type Extern and provide the ability to call functions, set globals, read memory, etc. When interacting with any wasm code you’ll want to make an Instance to call any code or execute anything.

Instances are owned by a Store which is passed in at creation time. It’s recommended to create instances with Linker::instantiate or similar Linker methods, but a more low-level constructor is also available as Instance::new.

Implementations

impl Instance

pub fn new(
store: impl AsContextMut,
module: &Module,
imports: &[Extern]
) -> Result<Instance, Error>

Creates a new Instance from the previously compiled Module and list of imports specified.

This method instantiates the module provided with the imports, following the procedure in the core specification to instantiate. Instantiation can fail for a number of reasons (many specified below), but if successful the start function will be automatically run (if specified in the module) and then the Instance will be returned.

Per the WebAssembly spec, instantiation includes running the module’s start function, if it has one (not to be confused with the _start function, which is not run).

Note that this is a low-level function that just performs an instantiation. See the Linker struct for an API which provides a convenient way to link imports and provides automatic Command and Reactor behavior.

Providing Imports

The entries in the list of imports are intended to correspond 1:1 with the list of imports returned by Module::imports. Before calling Instance::new you’ll want to inspect the return value of Module::imports and, for each import type, create an Extern which corresponds to that type. These Extern values are all then collected into a list and passed to this function.

Note that this function is intentionally relatively low level. For an easier time passing imports by doing name-based resolution it’s recommended to instead use the Linker type.

Errors

This function can fail for a number of reasons, including, but not limited to:

The number of imports provided doesn’t match the number of imports returned by the module’s Module::imports method.

The type of any Extern doesn’t match the corresponding ExternType entry that it maps to.

The start function in the instance, if present, traps.

Module/instance resource limits are exceeded.

When instantiation fails it’s recommended to inspect the return value to see why it failed, or bubble it upwards. If you’d like to specifically check for trap errors, you can use error.downcast::<Trap>().

Panics

This function will panic if called with a store associated with a asynchronous config. This function will also panic if any Extern supplied is not owned by store.

Examples found in repository

examples/gcd.rs (line 15)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22

fn main() -> Result<()> {
  // Load our WebAssembly (parsed WAT in our case), and then load it into a
  // `Module` which is attached to a `Store` cache. After we've got that we
  // can instantiate it.
  let mut store = Store::<()>::default();
  let module = Module::from_file(store.engine(), "examples/gcd.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `gcd` export
  let gcd = instance.get_typed_func::<(i32, i32), i32, _>(&mut store, "gcd")?;

  println!("gcd(6, 27) = {}", gcd.call(&mut store, (6, 27))?);
  Ok(())
}

examples/fib-debug/main.rs (line 21)

↕

14
15
16
17
18
19
20
21
22
23
24
25
26
27

fn main() -> Result<()> {
  // Load our previously compiled wasm file (built previously with Cargo) and
  // also ensure that we generate debuginfo so this executable can be
  // debugged in GDB.
  let engine = Engine::new(Config::new().debug_info(true))?;
  let mut store = Store::new(&engine, ());
  let module = Module::from_file(&engine, "target/wasm32-unknown-unknown/debug/fib.wasm")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fib` export
  let fib = instance.get_typed_func::<i32, i32, _>(&mut store, "fib")?;
  println!("fib(6) = {}", fib.call(&mut store, 6)?);
  Ok(())
}

examples/fuel.rs (line 15)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

fn main() -> Result<()> {
  let mut config = Config::new();
  config.consume_fuel(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());
  store.add_fuel(10_000)?;
  let module = Module::from_file(store.engine(), "examples/fuel.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fibonacci` export with higher and higher numbers until we exhaust our fuel.
  let fibonacci = instance.get_typed_func::<i32, i32, _>(&mut store, "fibonacci")?;
  for n in 1.. {
    let fuel_before = store.fuel_consumed().unwrap();
    let output = match fibonacci.call(&mut store, n) {
      Ok(v) => v,
      Err(_) => {
        println!("Exhausted fuel computing fib({})", n);
        break;
      }
    };
    let fuel_consumed = store.fuel_consumed().unwrap() - fuel_before;
    println!("fib({}) = {} [consumed {} fuel]", n, output, fuel_consumed);
    store.add_fuel(fuel_consumed)?;
  }
  Ok(())
}

examples/interrupt.rs (line 18)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

fn main() -> Result<()> {
  // Enable interruptable code via `Config` and then create an interrupt
  // handle which we'll use later to interrupt running code.
  let engine = Engine::new(Config::new().interruptable(true))?;
  let mut store = Store::new(&engine, ());
  let interrupt_handle = store.interrupt_handle()?;

  // Compile and instantiate a small example with an infinite loop.
  let module = Module::from_file(&engine, "examples/interrupt.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // Spin up a thread to send us an interrupt in a second
  std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    println!("Interrupting!");
    interrupt_handle.interrupt();
  });

  println!("Entering infinite loop ...");
  let trap = run.call(&mut store, ()).unwrap_err();

  println!("trap received...");
  assert!(trap.to_string().contains("wasm trap: interrupt"));

  Ok(())
}

examples/multi.rs (line 32)

↕

12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64

fn main() -> Result<()> {
  use wasmtime::*;

  println!("Initializing...");
  let engine = Engine::default();
  let mut store = Store::new(&engine, ());

  // Compile.
  println!("Compiling module...");
  let module = Module::from_file(&engine, "examples/multi.wat")?;

  // Create a host function which takes multiple parameters and returns
  // multiple results.
  println!("Creating callback...");
  let callback_func = Func::wrap(&mut store, |a: i32, b: i64| -> (i64, i32) {
    (b + 1, a + 1)
  });

  // Instantiate.
  println!("Instantiating module...");
  let instance = Instance::new(&mut store, &module, &[callback_func.into()])?;

  // Extract exports.
  println!("Extracting export...");
  let g = instance.get_typed_func::<(i32, i64), (i64, i32), _>(&mut store, "g")?;

  // Call `$g`.
  println!("Calling export \"g\"...");
  let (a, b) = g.call(&mut store, (1, 3))?;

  println!("Printing result...");
  println!("> {} {}", a, b);

  assert_eq!(a, 4);
  assert_eq!(b, 2);

  // Call `$round_trip_many`.
  println!("Calling export \"round_trip_many\"...");
  let round_trip_many = instance
        .get_typed_func::<
        (i64, i64, i64, i64, i64, i64, i64, i64, i64, i64),
        (i64, i64, i64, i64, i64, i64, i64, i64, i64, i64),
        _,
        >
        (&mut store, "round_trip_many")?;
  let results = round_trip_many.call(&mut store, (0, 1, 2, 3, 4, 5, 6, 7, 8, 9))?;

  println!("Printing result...");
  println!("> {:?}", results);
  assert_eq!(results, (0, 1, 2, 3, 4, 5, 6, 7, 8, 9));

  Ok(())
}

examples/serialize.rs (line 52)

↕

25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64

fn deserialize(buffer: &[u8]) -> Result<()> {
  // Configure the initial compilation environment, creating the global
  // `Store` structure. Note that you can also tweak configuration settings
  // with a `Config` and an `Engine` if desired.
  println!("Initializing...");
  let mut store: Store<()> = Store::default();

  // Compile the wasm binary into an in-memory instance of a `Module`. Note
  // that this is `unsafe` because it is our responsibility for guaranteeing
  // that these bytes are valid precompiled module bytes. We know that from
  // the structure of this example program.
  println!("Deserialize module...");
  let module = unsafe { Module::deserialize(store.engine(), buffer)? };

  // Here we handle the imports of the module, which in this case is our
  // `HelloCallback` type and its associated implementation of `Callback.
  println!("Creating callback...");
  let hello_func = Func::wrap(&mut store, || {
    println!("Calling back...");
    println!("> Hello World!");
  });

  // Once we've got that all set up we can then move to the instantiation
  // phase, pairing together a compiled module as well as a set of imports.
  // Note that this is where the wasm `start` function, if any, would run.
  println!("Instantiating module...");
  let imports = [hello_func.into()];
  let instance = Instance::new(&mut store, &module, &imports)?;

  // Next we poke around a bit to extract the `run` function from the module.
  println!("Extracting export...");
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // And last but not least we can call it!
  println!("Calling export...");
  run.call(&mut store, ())?;

  println!("Done.");
  Ok(())
}

Additional examples can be found in:

examples/externref.rs

examples/hello.rs

examples/memory.rs

examples/multimemory.rs

pub async fn new_async<T>(
store: impl AsContextMut<Data = T>,
module: &'_ Module,
imports: &'_ [Extern]
) -> Result<Instance, Error> where
T: Send,

Same as Instance::new, except for usage in [asynchronous stores].

For more details about this function see the documentation on Instance::new. The only difference between these two methods is that this one will asynchronously invoke the wasm start function in case it calls any imported function which is an asynchronous host function (e.g. created with Func::new_async.

Panics

This function will panic if called with a store associated with a synchronous config. This is only compatible with stores associated with an asynchronous config.

This function will also panic, like Instance::new, if any Extern specified does not belong to store.

pub fn ty(&self, store: impl AsContext) -> InstanceType

Returns the type signature of this instance.

Panics

Panics if store does not own this instance.

pub fn exports<'a, T>(
&'a self,
store: impl Into<StoreContextMut<'a, T>>
) -> impl ExactSizeIterator + 'a where
T: 'a,

Returns the list of exported items from this Instance.

Panics

Panics if store does not own this instance.

pub fn get_export(&self, store: impl AsContextMut, name: &str) -> Option<Extern>

Looks up an exported Extern value by name.

This method will search the module for an export named name and return the value, if found.

Returns None if there was no export named name.

Panics

Panics if store does not own this instance.

Why does get_export take a mutable context?

This method requires a mutable context because an instance’s exports are lazily populated, and we cache them as they are accessed. This makes instantiating a module faster, but also means this method requires a mutable context.

pub fn get_func(&self, store: impl AsContextMut, name: &str) -> Option<Func>

Looks up an exported Func value by name.

Returns None if there was no export named name, or if there was but it wasn’t a function.

Panics

Panics if store does not own this instance.

pub fn get_typed_func<Params, Results, S>(
&self,
store: S,
name: &str
) -> Result<TypedFunc<Params, Results>, Error> where
Params: WasmParams,
Results: WasmResults,
S: AsContextMut,

Looks up an exported Func value by name and with its type.

This function is a convenience wrapper over Instance::get_func and Func::typed. For more information see the linked documentation.

Returns an error if name isn’t a function export or if the export’s type did not match Params or Results

Panics

Panics if store does not own this instance.

Examples found in repository

examples/gcd.rs (line 18)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22

fn main() -> Result<()> {
  // Load our WebAssembly (parsed WAT in our case), and then load it into a
  // `Module` which is attached to a `Store` cache. After we've got that we
  // can instantiate it.
  let mut store = Store::<()>::default();
  let module = Module::from_file(store.engine(), "examples/gcd.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `gcd` export
  let gcd = instance.get_typed_func::<(i32, i32), i32, _>(&mut store, "gcd")?;

  println!("gcd(6, 27) = {}", gcd.call(&mut store, (6, 27))?);
  Ok(())
}

examples/fib-debug/main.rs (line 24)

↕

14
15
16
17
18
19
20
21
22
23
24
25
26
27

fn main() -> Result<()> {
  // Load our previously compiled wasm file (built previously with Cargo) and
  // also ensure that we generate debuginfo so this executable can be
  // debugged in GDB.
  let engine = Engine::new(Config::new().debug_info(true))?;
  let mut store = Store::new(&engine, ());
  let module = Module::from_file(&engine, "target/wasm32-unknown-unknown/debug/fib.wasm")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fib` export
  let fib = instance.get_typed_func::<i32, i32, _>(&mut store, "fib")?;
  println!("fib(6) = {}", fib.call(&mut store, 6)?);
  Ok(())
}

examples/threads.rs (line 55)

↕

49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70

fn run(engine: &Engine, module: &Module, linker: &Linker<()>) -> Result<()> {
  // Each sub-thread we have starting out by instantiating the `module`
  // provided into a fresh `Store`.
  println!("Instantiating module...");
  let mut store = Store::new(&engine, ());
  let instance = linker.instantiate(&mut store, module)?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  println!("Executing...");
  for _ in 0..N_REPS {
    run.call(&mut store, ())?;
    thread::sleep(time::Duration::from_millis(100));
  }

  // Also note that that a `Store` can also move between threads:
  println!("> Moving {:?} to a new thread", thread::current().id());
  let child = thread::spawn(move || run.call(&mut store, ()));

  child.join().unwrap()?;

  Ok(())
}

examples/tokio/main.rs (line 109)

↕

88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114

async fn run_wasm(inputs: Inputs) -> Result<(), Error> {
  let wasi = WasiCtxBuilder::new()
    // Let wasi print to this process's stdout.
    .inherit_stdout()
    // Set an environment variable so the wasm knows its name.
    .env("NAME", &inputs.name)?
    .build();
  let mut store = Store::new(&inputs.env.engine, wasi);

  // WebAssembly execution will be paused for an async yield every time it
  // consumes 10000 fuel. Fuel will be refilled u64::MAX times.
  store.out_of_fuel_async_yield(u64::MAX, 10000);

  // Instantiate into our own unique store using the shared linker, afterwards
  // acquiring the `_start` function for the module and executing it.
  let instance = inputs
    .env
    .linker
    .instantiate_async(&mut store, &inputs.env.module)
    .await?;
  instance
    .get_typed_func::<(), (), _>(&mut store, "_start")?
    .call_async(&mut store, ())
    .await?;

  Ok(())
}

examples/fuel.rs (line 18)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

fn main() -> Result<()> {
  let mut config = Config::new();
  config.consume_fuel(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());
  store.add_fuel(10_000)?;
  let module = Module::from_file(store.engine(), "examples/fuel.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fibonacci` export with higher and higher numbers until we exhaust our fuel.
  let fibonacci = instance.get_typed_func::<i32, i32, _>(&mut store, "fibonacci")?;
  for n in 1.. {
    let fuel_before = store.fuel_consumed().unwrap();
    let output = match fibonacci.call(&mut store, n) {
      Ok(v) => v,
      Err(_) => {
        println!("Exhausted fuel computing fib({})", n);
        break;
      }
    };
    let fuel_consumed = store.fuel_consumed().unwrap() - fuel_before;
    println!("fib({}) = {} [consumed {} fuel]", n, output, fuel_consumed);
    store.add_fuel(fuel_consumed)?;
  }
  Ok(())
}

examples/interrupt.rs (line 19)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

fn main() -> Result<()> {
  // Enable interruptable code via `Config` and then create an interrupt
  // handle which we'll use later to interrupt running code.
  let engine = Engine::new(Config::new().interruptable(true))?;
  let mut store = Store::new(&engine, ());
  let interrupt_handle = store.interrupt_handle()?;

  // Compile and instantiate a small example with an infinite loop.
  let module = Module::from_file(&engine, "examples/interrupt.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // Spin up a thread to send us an interrupt in a second
  std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    println!("Interrupting!");
    interrupt_handle.interrupt();
  });

  println!("Entering infinite loop ...");
  let trap = run.call(&mut store, ()).unwrap_err();

  println!("trap received...");
  assert!(trap.to_string().contains("wasm trap: interrupt"));

  Ok(())
}

Additional examples can be found in:

examples/linking.rs

examples/multi.rs

examples/serialize.rs

examples/externref.rs

examples/hello.rs

examples/memory.rs

examples/multimemory.rs

pub fn get_table(&self, store: impl AsContextMut, name: &str) -> Option<Table>

Looks up an exported Table value by name.

Returns None if there was no export named name, or if there was but it wasn’t a table.

Panics

Panics if store does not own this instance.

Examples found in repository

examples/externref.rs (line 30)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57

fn main() -> Result<()> {
  println!("Initializing...");
  let mut config = Config::new();
  config.wasm_reference_types(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());

  println!("Compiling module...");
  let module = Module::from_file(&engine, "examples/externref.wat")?;

  println!("Instantiating module...");
  let instance = Instance::new(&mut store, &module, &[])?;

  println!("Creating new `externref`...");
  let externref = ExternRef::new("Hello, World!");
  assert!(externref.data().is::<&'static str>());
  assert_eq!(
    *externref.data().downcast_ref::<&'static str>().unwrap(),
    "Hello, World!"
  );

  println!("Touching `externref` table...");
  let table = instance.get_table(&mut store, "table").unwrap();
  table.set(&mut store, 3, Some(externref.clone()).into())?;
  let elem = table
    .get(&mut store, 3)
    .unwrap() // assert in bounds
    .unwrap_externref() // assert it's an externref table
    .unwrap(); // assert the externref isn't null
  assert!(elem.ptr_eq(&externref));

  println!("Touching `externref` global...");
  let global = instance.get_global(&mut store, "global").unwrap();
  global.set(&mut store, Some(externref.clone()).into())?;
  let global_val = global.get(&mut store).unwrap_externref().unwrap();
  assert!(global_val.ptr_eq(&externref));

  println!("Calling `externref` func...");
  let func =
    instance.get_typed_func::<Option<ExternRef>, Option<ExternRef>, _>(&mut store, "func")?;
  let ret = func.call(&mut store, Some(externref.clone()))?;
  assert!(ret.is_some());
  assert!(ret.unwrap().ptr_eq(&externref));

  println!("GCing within the store...");
  store.gc();

  println!("Done.");
  Ok(())
}

pub fn get_memory(&self, store: impl AsContextMut, name: &str) -> Option<Memory>

Looks up an exported Memory value by name.

Returns None if there was no export named name, or if there was but it wasn’t a memory.

Panics

Panics if store does not own this instance.

Examples found in repository

examples/memory.rs (line 21)

↕

12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74

fn main() -> Result<()> {
  // Create our `store_fn` context and then compile a module and create an
  // instance from the compiled module all in one go.
  let mut store: Store<()> = Store::default();
  let module = Module::from_file(store.engine(), "examples/memory.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // load_fn up our exports from the instance
  let memory = instance
    .get_memory(&mut store, "memory")
    .ok_or(anyhow::format_err!("failed to find `memory` export"))?;
  let size = instance.get_typed_func::<(), i32, _>(&mut store, "size")?;
  let load_fn = instance.get_typed_func::<i32, i32, _>(&mut store, "load")?;
  let store_fn = instance.get_typed_func::<(i32, i32), (), _>(&mut store, "store")?;

  println!("Checking memory...");
  assert_eq!(memory.size(&store), 2);
  assert_eq!(memory.data_size(&store), 0x20000);
  assert_eq!(memory.data_mut(&mut store)[0], 0);
  assert_eq!(memory.data_mut(&mut store)[0x1000], 1);
  assert_eq!(memory.data_mut(&mut store)[0x1003], 4);

  assert_eq!(size.call(&mut store, ())?, 2);
  assert_eq!(load_fn.call(&mut store, 0)?, 0);
  assert_eq!(load_fn.call(&mut store, 0x1000)?, 1);
  assert_eq!(load_fn.call(&mut store, 0x1003)?, 4);
  assert_eq!(load_fn.call(&mut store, 0x1ffff)?, 0);
  assert!(load_fn.call(&mut store, 0x20000).is_err()); // out of bounds trap

  println!("Mutating memory...");
  memory.data_mut(&mut store)[0x1003] = 5;

  store_fn.call(&mut store, (0x1002, 6))?;
  assert!(store_fn.call(&mut store, (0x20000, 0)).is_err()); // out of bounds trap

  assert_eq!(memory.data(&store)[0x1002], 6);
  assert_eq!(memory.data(&store)[0x1003], 5);
  assert_eq!(load_fn.call(&mut store, 0x1002)?, 6);
  assert_eq!(load_fn.call(&mut store, 0x1003)?, 5);

  // Grow memory.
  println!("Growing memory...");
  memory.grow(&mut store, 1)?;
  assert_eq!(memory.size(&store), 3);
  assert_eq!(memory.data_size(&store), 0x30000);

  assert_eq!(load_fn.call(&mut store, 0x20000)?, 0);
  store_fn.call(&mut store, (0x20000, 0))?;
  assert!(load_fn.call(&mut store, 0x30000).is_err());
  assert!(store_fn.call(&mut store, (0x30000, 0)).is_err());

  assert!(memory.grow(&mut store, 1).is_err());
  assert!(memory.grow(&mut store, 0).is_ok());

  println!("Creating stand-alone memory...");
  let memorytype = MemoryType::new(5, Some(5));
  let memory2 = Memory::new(&mut store, memorytype)?;
  assert_eq!(memory2.size(&store), 5);
  assert!(memory2.grow(&mut store, 1).is_err());
  assert!(memory2.grow(&mut store, 0).is_ok());

  Ok(())
}

examples/multimemory.rs (line 26)

≺ ≻↕

 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121

fn main() -> Result<()> {
  // Enable the multi-memory feature.
  let mut config = Config::new();
  config.wasm_multi_memory(true);

  let engine = Engine::new(&config)?;

  // Create our `store_fn` context and then compile a module and create an
  // instance from the compiled module all in one go.
  let mut store = Store::new(&engine, ());
  let module = Module::from_file(store.engine(), "examples/multimemory.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  let memory0 = instance
    .get_memory(&mut store, "memory0")
    .ok_or(anyhow::format_err!("failed to find `memory0` export"))?;
  let size0 = instance.get_typed_func::<(), i32, _>(&mut store, "size0")?;
  let load0 = instance.get_typed_func::<i32, i32, _>(&mut store, "load0")?;
  let store0 = instance.get_typed_func::<(i32, i32), (), _>(&mut store, "store0")?;

  let memory1 = instance
    .get_memory(&mut store, "memory1")
    .ok_or(anyhow::format_err!("failed to find `memory1` export"))?;
  let size1 = instance.get_typed_func::<(), i32, _>(&mut store, "size1")?;
  let load1 = instance.get_typed_func::<i32, i32, _>(&mut store, "load1")?;
  let store1 = instance.get_typed_func::<(i32, i32), (), _>(&mut store, "store1")?;

  println!("Checking memory...");
  assert_eq!(memory0.size(&store), 2);
  assert_eq!(memory0.data_size(&store), 0x20000);
  assert_eq!(memory0.data_mut(&mut store)[0], 0);
  assert_eq!(memory0.data_mut(&mut store)[0x1000], 1);
  assert_eq!(memory0.data_mut(&mut store)[0x1001], 2);
  assert_eq!(memory0.data_mut(&mut store)[0x1002], 3);
  assert_eq!(memory0.data_mut(&mut store)[0x1003], 4);

  assert_eq!(size0.call(&mut store, ())?, 2);
  assert_eq!(load0.call(&mut store, 0)?, 0);
  assert_eq!(load0.call(&mut store, 0x1000)?, 1);
  assert_eq!(load0.call(&mut store, 0x1001)?, 2);
  assert_eq!(load0.call(&mut store, 0x1002)?, 3);
  assert_eq!(load0.call(&mut store, 0x1003)?, 4);
  assert_eq!(load0.call(&mut store, 0x1ffff)?, 0);
  assert!(load0.call(&mut store, 0x20000).is_err()); // out of bounds trap

  assert_eq!(memory1.size(&store), 2);
  assert_eq!(memory1.data_size(&store), 0x20000);
  assert_eq!(memory1.data_mut(&mut store)[0], 0);
  assert_eq!(memory1.data_mut(&mut store)[0x1000], 4);
  assert_eq!(memory1.data_mut(&mut store)[0x1001], 3);
  assert_eq!(memory1.data_mut(&mut store)[0x1002], 2);
  assert_eq!(memory1.data_mut(&mut store)[0x1003], 1);

  assert_eq!(size1.call(&mut store, ())?, 2);
  assert_eq!(load1.call(&mut store, 0)?, 0);
  assert_eq!(load1.call(&mut store, 0x1000)?, 4);
  assert_eq!(load1.call(&mut store, 0x1001)?, 3);
  assert_eq!(load1.call(&mut store, 0x1002)?, 2);
  assert_eq!(load1.call(&mut store, 0x1003)?, 1);
  assert_eq!(load1.call(&mut store, 0x1ffff)?, 0);
  assert!(load0.call(&mut store, 0x20000).is_err()); // out of bounds trap

  println!("Mutating memory...");
  memory0.data_mut(&mut store)[0x1003] = 5;

  store0.call(&mut store, (0x1002, 6))?;
  assert!(store0.call(&mut store, (0x20000, 0)).is_err()); // out of bounds trap

  assert_eq!(memory0.data(&store)[0x1002], 6);
  assert_eq!(memory0.data(&store)[0x1003], 5);
  assert_eq!(load0.call(&mut store, 0x1002)?, 6);
  assert_eq!(load0.call(&mut store, 0x1003)?, 5);

  memory1.data_mut(&mut store)[0x1003] = 7;

  store1.call(&mut store, (0x1002, 8))?;
  assert!(store1.call(&mut store, (0x20000, 0)).is_err()); // out of bounds trap

  assert_eq!(memory1.data(&store)[0x1002], 8);
  assert_eq!(memory1.data(&store)[0x1003], 7);
  assert_eq!(load1.call(&mut store, 0x1002)?, 8);
  assert_eq!(load1.call(&mut store, 0x1003)?, 7);

  println!("Growing memory...");
  memory0.grow(&mut store, 1)?;
  assert_eq!(memory0.size(&store), 3);
  assert_eq!(memory0.data_size(&store), 0x30000);

  assert_eq!(load0.call(&mut store, 0x20000)?, 0);
  store0.call(&mut store, (0x20000, 0))?;
  assert!(load0.call(&mut store, 0x30000).is_err());
  assert!(store0.call(&mut store, (0x30000, 0)).is_err());

  assert!(memory0.grow(&mut store, 1).is_err());
  assert!(memory0.grow(&mut store, 0).is_ok());

  memory1.grow(&mut store, 2)?;
  assert_eq!(memory1.size(&store), 4);
  assert_eq!(memory1.data_size(&store), 0x40000);

  assert_eq!(load1.call(&mut store, 0x30000)?, 0);
  store1.call(&mut store, (0x30000, 0))?;
  assert!(load1.call(&mut store, 0x40000).is_err());
  assert!(store1.call(&mut store, (0x40000, 0)).is_err());

  assert!(memory1.grow(&mut store, 1).is_err());
  assert!(memory1.grow(&mut store, 0).is_ok());

  Ok(())
}

pub fn get_global(&self, store: impl AsContextMut, name: &str) -> Option<Global>

Looks up an exported Global value by name.

Returns None if there was no export named name, or if there was but it wasn’t a global.

Panics

Panics if store does not own this instance.

Examples found in repository

examples/externref.rs (line 40)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57

fn main() -> Result<()> {
  println!("Initializing...");
  let mut config = Config::new();
  config.wasm_reference_types(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());

  println!("Compiling module...");
  let module = Module::from_file(&engine, "examples/externref.wat")?;

  println!("Instantiating module...");
  let instance = Instance::new(&mut store, &module, &[])?;

  println!("Creating new `externref`...");
  let externref = ExternRef::new("Hello, World!");
  assert!(externref.data().is::<&'static str>());
  assert_eq!(
    *externref.data().downcast_ref::<&'static str>().unwrap(),
    "Hello, World!"
  );

  println!("Touching `externref` table...");
  let table = instance.get_table(&mut store, "table").unwrap();
  table.set(&mut store, 3, Some(externref.clone()).into())?;
  let elem = table
    .get(&mut store, 3)
    .unwrap() // assert in bounds
    .unwrap_externref() // assert it's an externref table
    .unwrap(); // assert the externref isn't null
  assert!(elem.ptr_eq(&externref));

  println!("Touching `externref` global...");
  let global = instance.get_global(&mut store, "global").unwrap();
  global.set(&mut store, Some(externref.clone()).into())?;
  let global_val = global.get(&mut store).unwrap_externref().unwrap();
  assert!(global_val.ptr_eq(&externref));

  println!("Calling `externref` func...");
  let func =
    instance.get_typed_func::<Option<ExternRef>, Option<ExternRef>, _>(&mut store, "func")?;
  let ret = func.call(&mut store, Some(externref.clone()))?;
  assert!(ret.is_some());
  assert!(ret.unwrap().ptr_eq(&externref));

  println!("GCing within the store...");
  store.gc();

  println!("Done.");
  Ok(())
}