Struct wiggle::wasmtime_crate::InterruptHandle

pub struct InterruptHandle { /* fields omitted */ }

A threadsafe handle used to interrupt instances executing within a particular Store.

This structure is created by the Store::interrupt_handle method.

Implementations

impl InterruptHandle

pub fn interrupt(&self)

Flags that execution within this handle’s original Store should be interrupted.

This will not immediately interrupt execution of wasm modules, but rather it will interrupt wasm execution of loop headers and wasm execution of function entries. For more information see Store::interrupt_handle.

Examples found in repository

examples/interrupt.rs (line 25)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

fn main() -> Result<()> {
  // Enable interruptable code via `Config` and then create an interrupt
  // handle which we'll use later to interrupt running code.
  let engine = Engine::new(Config::new().interruptable(true))?;
  let mut store = Store::new(&engine, ());
  let interrupt_handle = store.interrupt_handle()?;

  // Compile and instantiate a small example with an infinite loop.
  let module = Module::from_file(&engine, "examples/interrupt.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // Spin up a thread to send us an interrupt in a second
  std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    println!("Interrupting!");
    interrupt_handle.interrupt();
  });

  println!("Entering infinite loop ...");
  let trap = run.call(&mut store, ()).unwrap_err();

  println!("trap received...");
  assert!(trap.to_string().contains("wasm trap: interrupt"));

  Ok(())
}