Struct wiggle::wasmtime_crate::MemoryType

pub struct MemoryType { /* fields omitted */ }

A descriptor for a WebAssembly memory type.

Memories are described in units of pages (64KB) and represent contiguous chunks of addressable memory.

Implementations

impl MemoryType

pub fn new(minimum: u32, maximum: Option<u32>) -> MemoryType

Creates a new descriptor for a 32-bit WebAssembly memory given the specified limits of the memory.

The minimum and maximum values here are specified in units of WebAssembly pages, which are 64k.

Examples found in repository

examples/memory.rs (line 67)

↕

12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74

fn main() -> Result<()> {
  // Create our `store_fn` context and then compile a module and create an
  // instance from the compiled module all in one go.
  let mut store: Store<()> = Store::default();
  let module = Module::from_file(store.engine(), "examples/memory.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // load_fn up our exports from the instance
  let memory = instance
    .get_memory(&mut store, "memory")
    .ok_or(anyhow::format_err!("failed to find `memory` export"))?;
  let size = instance.get_typed_func::<(), i32, _>(&mut store, "size")?;
  let load_fn = instance.get_typed_func::<i32, i32, _>(&mut store, "load")?;
  let store_fn = instance.get_typed_func::<(i32, i32), (), _>(&mut store, "store")?;

  println!("Checking memory...");
  assert_eq!(memory.size(&store), 2);
  assert_eq!(memory.data_size(&store), 0x20000);
  assert_eq!(memory.data_mut(&mut store)[0], 0);
  assert_eq!(memory.data_mut(&mut store)[0x1000], 1);
  assert_eq!(memory.data_mut(&mut store)[0x1003], 4);

  assert_eq!(size.call(&mut store, ())?, 2);
  assert_eq!(load_fn.call(&mut store, 0)?, 0);
  assert_eq!(load_fn.call(&mut store, 0x1000)?, 1);
  assert_eq!(load_fn.call(&mut store, 0x1003)?, 4);
  assert_eq!(load_fn.call(&mut store, 0x1ffff)?, 0);
  assert!(load_fn.call(&mut store, 0x20000).is_err()); // out of bounds trap

  println!("Mutating memory...");
  memory.data_mut(&mut store)[0x1003] = 5;

  store_fn.call(&mut store, (0x1002, 6))?;
  assert!(store_fn.call(&mut store, (0x20000, 0)).is_err()); // out of bounds trap

  assert_eq!(memory.data(&store)[0x1002], 6);
  assert_eq!(memory.data(&store)[0x1003], 5);
  assert_eq!(load_fn.call(&mut store, 0x1002)?, 6);
  assert_eq!(load_fn.call(&mut store, 0x1003)?, 5);

  // Grow memory.
  println!("Growing memory...");
  memory.grow(&mut store, 1)?;
  assert_eq!(memory.size(&store), 3);
  assert_eq!(memory.data_size(&store), 0x30000);

  assert_eq!(load_fn.call(&mut store, 0x20000)?, 0);
  store_fn.call(&mut store, (0x20000, 0))?;
  assert!(load_fn.call(&mut store, 0x30000).is_err());
  assert!(store_fn.call(&mut store, (0x30000, 0)).is_err());

  assert!(memory.grow(&mut store, 1).is_err());
  assert!(memory.grow(&mut store, 0).is_ok());

  println!("Creating stand-alone memory...");
  let memorytype = MemoryType::new(5, Some(5));
  let memory2 = Memory::new(&mut store, memorytype)?;
  assert_eq!(memory2.size(&store), 5);
  assert!(memory2.grow(&mut store, 1).is_err());
  assert!(memory2.grow(&mut store, 0).is_ok());

  Ok(())
}

pub fn new64(minimum: u64, maximum: Option<u64>) -> MemoryType

Creates a new descriptor for a 64-bit WebAssembly memory given the specified limits of the memory.

The minimum and maximum values here are specified in units of WebAssembly pages, which are 64k.

Note that 64-bit memories are part of the memory64 proposal for WebAssembly which is not standardized yet.

pub fn is_64(&self) -> bool

Returns whether this is a 64-bit memory or not.

Note that 64-bit memories are part of the memory64 proposal for WebAssembly which is not standardized yet.

pub fn minimum(&self) -> u64

Returns minimum number of WebAssembly pages this memory must have.

Note that the return value, while a u64, will always fit into a u32 for 32-bit memories.

pub fn maximum(&self) -> Option<u64>

Returns the optionally-specified maximum number of pages this memory can have.

If this returns None then the memory is not limited in size.

Note that the return value, while a u64, will always fit into a u32 for 32-bit memories.