Struct wiggle::wasmtime_crate::Store

pub struct Store<T> { /* fields omitted */ }

A Store is a collection of WebAssembly instances and host-defined state.

All WebAssembly instances and items will be attached to and refer to a Store. For example instances, functions, globals, and tables are all attached to a Store. Instances are created by instantiating a Module within a Store.

A Store is intended to be a short-lived object in a program. No form of GC is implemented at this time so once an instance is created within a Store it will not be deallocated until the Store itself is dropped. This makes Store unsuitable for creating an unbounded number of instances in it because Store will never release this memory. It’s recommended to have a Store correspond roughly to the lifetime of a “main instance” that an embedding is interested in executing.

Type parameter T

Each Store has a type parameter T associated with it. This T represents state defined by the host. This state will be accessible through the Caller type that host-defined functions get access to. This T is suitable for storing Store-specific information which imported functions may want access to.

The data T can be accessed through methods like Store::data and Store::data_mut.

Stores, contexts, oh my

Most methods in Wasmtime take something of the form AsContext or AsContextMut as the first argument. These two traits allow ergonomically passing in the context you currently have to any method. The primary two sources of contexts are:

Store<T>

Caller<'_, T>

corresponding to what you create and what you have access to in a host function. You can also explicitly acquire a StoreContext or StoreContextMut and pass that around as well.

Note that all methods on Store are mirrored onto StoreContext, StoreContextMut, and Caller. This way no matter what form of context you have you can call various methods, create objects, etc.

Stores and Default

You can create a store with default configuration settings using Store::default(). This will create a brand new Engine with default configuration (see Config for more information).

Implementations

impl<T> Store<T>

pub fn new(engine: &Engine, data: T) -> Store<T>

Creates a new Store to be associated with the given Engine and data provided.

The created Store will place no additional limits on the size of linear memories or tables at runtime. Linear memories and tables will be allowed to grow to any upper limit specified in their definitions. The store will limit the number of instances, linear memories, and tables created to 10,000. This can be overridden with the Store::limiter configuration method.

Examples found in repository

examples/fib-debug/main.rs (line 19)

↕

14
15
16
17
18
19
20
21
22
23
24
25
26
27

fn main() -> Result<()> {
  // Load our previously compiled wasm file (built previously with Cargo) and
  // also ensure that we generate debuginfo so this executable can be
  // debugged in GDB.
  let engine = Engine::new(Config::new().debug_info(true))?;
  let mut store = Store::new(&engine, ());
  let module = Module::from_file(&engine, "target/wasm32-unknown-unknown/debug/fib.wasm")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fib` export
  let fib = instance.get_typed_func::<i32, i32, _>(&mut store, "fib")?;
  println!("fib(6) = {}", fib.call(&mut store, 6)?);
  Ok(())
}

examples/threads.rs (line 53)

↕

49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70

fn run(engine: &Engine, module: &Module, linker: &Linker<()>) -> Result<()> {
  // Each sub-thread we have starting out by instantiating the `module`
  // provided into a fresh `Store`.
  println!("Instantiating module...");
  let mut store = Store::new(&engine, ());
  let instance = linker.instantiate(&mut store, module)?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  println!("Executing...");
  for _ in 0..N_REPS {
    run.call(&mut store, ())?;
    thread::sleep(time::Duration::from_millis(100));
  }

  // Also note that that a `Store` can also move between threads:
  println!("> Moving {:?} to a new thread", thread::current().id());
  let child = thread::spawn(move || run.call(&mut store, ()));

  child.join().unwrap()?;

  Ok(())
}

examples/wasi/main.rs (line 23)

↕

10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34

fn main() -> Result<()> {
  // Define the WASI functions globally on the `Config`.
  let engine = Engine::default();
  let mut linker = Linker::new(&engine);
  wasmtime_wasi::add_to_linker(&mut linker, |s| s)?;

  // Create a WASI context and put it in a Store; all instances in the store
  // share this context. `WasiCtxBuilder` provides a number of ways to
  // configure what the target program will have access to.
  let wasi = WasiCtxBuilder::new()
    .inherit_stdio()
    .inherit_args()?
    .build();
  let mut store = Store::new(&engine, wasi);

  // Instantiate our module with the imports we've created, and run it.
  let module = Module::from_file(&engine, "target/wasm32-wasi/debug/wasi.wasm")?;
  linker.module(&mut store, "", &module)?;
  linker
    .get_default(&mut store, "")?
    .typed::<(), (), _>(&store)?
    .call(&mut store, ())?;

  Ok(())
}

examples/tokio/main.rs (line 95)

↕

88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114

async fn run_wasm(inputs: Inputs) -> Result<(), Error> {
  let wasi = WasiCtxBuilder::new()
    // Let wasi print to this process's stdout.
    .inherit_stdout()
    // Set an environment variable so the wasm knows its name.
    .env("NAME", &inputs.name)?
    .build();
  let mut store = Store::new(&inputs.env.engine, wasi);

  // WebAssembly execution will be paused for an async yield every time it
  // consumes 10000 fuel. Fuel will be refilled u64::MAX times.
  store.out_of_fuel_async_yield(u64::MAX, 10000);

  // Instantiate into our own unique store using the shared linker, afterwards
  // acquiring the `_start` function for the module and executing it.
  let instance = inputs
    .env
    .linker
    .instantiate_async(&mut store, &inputs.env.module)
    .await?;
  instance
    .get_typed_func::<(), (), _>(&mut store, "_start")?
    .call_async(&mut store, ())
    .await?;

  Ok(())
}

examples/fuel.rs (line 12)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

fn main() -> Result<()> {
  let mut config = Config::new();
  config.consume_fuel(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());
  store.add_fuel(10_000)?;
  let module = Module::from_file(store.engine(), "examples/fuel.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fibonacci` export with higher and higher numbers until we exhaust our fuel.
  let fibonacci = instance.get_typed_func::<i32, i32, _>(&mut store, "fibonacci")?;
  for n in 1.. {
    let fuel_before = store.fuel_consumed().unwrap();
    let output = match fibonacci.call(&mut store, n) {
      Ok(v) => v,
      Err(_) => {
        println!("Exhausted fuel computing fib({})", n);
        break;
      }
    };
    let fuel_consumed = store.fuel_consumed().unwrap() - fuel_before;
    println!("fib({}) = {} [consumed {} fuel]", n, output, fuel_consumed);
    store.add_fuel(fuel_consumed)?;
  }
  Ok(())
}

examples/interrupt.rs (line 13)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

fn main() -> Result<()> {
  // Enable interruptable code via `Config` and then create an interrupt
  // handle which we'll use later to interrupt running code.
  let engine = Engine::new(Config::new().interruptable(true))?;
  let mut store = Store::new(&engine, ());
  let interrupt_handle = store.interrupt_handle()?;

  // Compile and instantiate a small example with an infinite loop.
  let module = Module::from_file(&engine, "examples/interrupt.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // Spin up a thread to send us an interrupt in a second
  std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    println!("Interrupting!");
    interrupt_handle.interrupt();
  });

  println!("Entering infinite loop ...");
  let trap = run.call(&mut store, ()).unwrap_err();

  println!("trap received...");
  assert!(trap.to_string().contains("wasm trap: interrupt"));

  Ok(())
}

Additional examples can be found in:

examples/linking.rs

examples/multi.rs

examples/externref.rs

examples/hello.rs

examples/multimemory.rs

pub fn data(&self) -> &T

Access the underlying data owned by this Store.

pub fn data_mut(&mut self) -> &mut T

Access the underlying data owned by this Store.

pub fn into_data(self) -> T

Consumes this Store, destroying it, and returns the underlying data.

pub fn limiter(
&mut self,
limiter: impl FnMut(&mut T) -> &mut dyn ResourceLimiter + Send + Sync + 'static
)

Configures the ResourceLimiter used to limit resource creation within this Store.

Note that this limiter is only used to limit the creation/growth of resources in the future, this does not retroactively attempt to apply limits to the Store.

pub fn limiter_async(
&mut self,
limiter: impl FnMut(&mut T) -> &mut dyn ResourceLimiterAsync + Send + Sync + 'static
)

Configures the ResourceLimiterAsync used to limit resource creation within this Store. Must be used with an async Store!.

Note that this limiter is only used to limit the creation/growth of resources in the future, this does not retroactively attempt to apply limits to the Store.

This variation on the ResourceLimiter makes the memory_growing and table_growing functions async. This means that, as part of your resource limiting strategy, the async resource limiter may yield execution until a resource becomes available.

By using a ResourceLimiterAsync with a Store, you can no longer use Memory::new, Memory::grow, Table::new, and Table::grow. Instead, you must use their async variants: Memory::new_async, Memory::grow_async, Table::new_async, and Table::grow_async.

pub fn call_hook(
&mut self,
hook: impl FnMut(&mut T, CallHook) -> Result<(), Trap> + Send + Sync + 'static
)

Configure a function that runs on calls and returns between WebAssembly and host code.

The function is passed a CallHook argument, which indicates which state transition the VM is making.

This function may return a Trap. If a trap is returned when an import was called, it is immediately raised as-if the host import had returned the trap. If a trap is returned after wasm returns to the host then the wasm function’s result is ignored and this trap is returned instead.

After this function returns a trap, it may be called for subsequent returns to host or wasm code as the trap propogates to the root call.

pub fn engine(&self) -> &Engine

Returns the Engine that this store is associated with.

Examples found in repository

examples/gcd.rs (line 14)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22

fn main() -> Result<()> {
  // Load our WebAssembly (parsed WAT in our case), and then load it into a
  // `Module` which is attached to a `Store` cache. After we've got that we
  // can instantiate it.
  let mut store = Store::<()>::default();
  let module = Module::from_file(store.engine(), "examples/gcd.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `gcd` export
  let gcd = instance.get_typed_func::<(i32, i32), i32, _>(&mut store, "gcd")?;

  println!("gcd(6, 27) = {}", gcd.call(&mut store, (6, 27))?);
  Ok(())
}

examples/fuel.rs (line 14)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

fn main() -> Result<()> {
  let mut config = Config::new();
  config.consume_fuel(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());
  store.add_fuel(10_000)?;
  let module = Module::from_file(store.engine(), "examples/fuel.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fibonacci` export with higher and higher numbers until we exhaust our fuel.
  let fibonacci = instance.get_typed_func::<i32, i32, _>(&mut store, "fibonacci")?;
  for n in 1.. {
    let fuel_before = store.fuel_consumed().unwrap();
    let output = match fibonacci.call(&mut store, n) {
      Ok(v) => v,
      Err(_) => {
        println!("Exhausted fuel computing fib({})", n);
        break;
      }
    };
    let fuel_consumed = store.fuel_consumed().unwrap() - fuel_before;
    println!("fib({}) = {} [consumed {} fuel]", n, output, fuel_consumed);
    store.add_fuel(fuel_consumed)?;
  }
  Ok(())
}

examples/serialize.rs (line 37)

↕

25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64

fn deserialize(buffer: &[u8]) -> Result<()> {
  // Configure the initial compilation environment, creating the global
  // `Store` structure. Note that you can also tweak configuration settings
  // with a `Config` and an `Engine` if desired.
  println!("Initializing...");
  let mut store: Store<()> = Store::default();

  // Compile the wasm binary into an in-memory instance of a `Module`. Note
  // that this is `unsafe` because it is our responsibility for guaranteeing
  // that these bytes are valid precompiled module bytes. We know that from
  // the structure of this example program.
  println!("Deserialize module...");
  let module = unsafe { Module::deserialize(store.engine(), buffer)? };

  // Here we handle the imports of the module, which in this case is our
  // `HelloCallback` type and its associated implementation of `Callback.
  println!("Creating callback...");
  let hello_func = Func::wrap(&mut store, || {
    println!("Calling back...");
    println!("> Hello World!");
  });

  // Once we've got that all set up we can then move to the instantiation
  // phase, pairing together a compiled module as well as a set of imports.
  // Note that this is where the wasm `start` function, if any, would run.
  println!("Instantiating module...");
  let imports = [hello_func.into()];
  let instance = Instance::new(&mut store, &module, &imports)?;

  // Next we poke around a bit to extract the `run` function from the module.
  println!("Extracting export...");
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // And last but not least we can call it!
  println!("Calling export...");
  run.call(&mut store, ())?;

  println!("Done.");
  Ok(())
}

examples/memory.rs (line 16)

↕

12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74

fn main() -> Result<()> {
  // Create our `store_fn` context and then compile a module and create an
  // instance from the compiled module all in one go.
  let mut store: Store<()> = Store::default();
  let module = Module::from_file(store.engine(), "examples/memory.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // load_fn up our exports from the instance
  let memory = instance
    .get_memory(&mut store, "memory")
    .ok_or(anyhow::format_err!("failed to find `memory` export"))?;
  let size = instance.get_typed_func::<(), i32, _>(&mut store, "size")?;
  let load_fn = instance.get_typed_func::<i32, i32, _>(&mut store, "load")?;
  let store_fn = instance.get_typed_func::<(i32, i32), (), _>(&mut store, "store")?;

  println!("Checking memory...");
  assert_eq!(memory.size(&store), 2);
  assert_eq!(memory.data_size(&store), 0x20000);
  assert_eq!(memory.data_mut(&mut store)[0], 0);
  assert_eq!(memory.data_mut(&mut store)[0x1000], 1);
  assert_eq!(memory.data_mut(&mut store)[0x1003], 4);

  assert_eq!(size.call(&mut store, ())?, 2);
  assert_eq!(load_fn.call(&mut store, 0)?, 0);
  assert_eq!(load_fn.call(&mut store, 0x1000)?, 1);
  assert_eq!(load_fn.call(&mut store, 0x1003)?, 4);
  assert_eq!(load_fn.call(&mut store, 0x1ffff)?, 0);
  assert!(load_fn.call(&mut store, 0x20000).is_err()); // out of bounds trap

  println!("Mutating memory...");
  memory.data_mut(&mut store)[0x1003] = 5;

  store_fn.call(&mut store, (0x1002, 6))?;
  assert!(store_fn.call(&mut store, (0x20000, 0)).is_err()); // out of bounds trap

  assert_eq!(memory.data(&store)[0x1002], 6);
  assert_eq!(memory.data(&store)[0x1003], 5);
  assert_eq!(load_fn.call(&mut store, 0x1002)?, 6);
  assert_eq!(load_fn.call(&mut store, 0x1003)?, 5);

  // Grow memory.
  println!("Growing memory...");
  memory.grow(&mut store, 1)?;
  assert_eq!(memory.size(&store), 3);
  assert_eq!(memory.data_size(&store), 0x30000);

  assert_eq!(load_fn.call(&mut store, 0x20000)?, 0);
  store_fn.call(&mut store, (0x20000, 0))?;
  assert!(load_fn.call(&mut store, 0x30000).is_err());
  assert!(store_fn.call(&mut store, (0x30000, 0)).is_err());

  assert!(memory.grow(&mut store, 1).is_err());
  assert!(memory.grow(&mut store, 0).is_ok());

  println!("Creating stand-alone memory...");
  let memorytype = MemoryType::new(5, Some(5));
  let memory2 = Memory::new(&mut store, memorytype)?;
  assert_eq!(memory2.size(&store), 5);
  assert!(memory2.grow(&mut store, 1).is_err());
  assert!(memory2.grow(&mut store, 0).is_ok());

  Ok(())
}

examples/multimemory.rs (line 22)

↕

 12
 13
 14
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24
 25
 26
 27
 28
 29
 30
 31
 32
 33
 34
 35
 36
 37
 38
 39
 40
 41
 42
 43
 44
 45
 46
 47
 48
 49
 50
 51
 52
 53
 54
 55
 56
 57
 58
 59
 60
 61
 62
 63
 64
 65
 66
 67
 68
 69
 70
 71
 72
 73
 74
 75
 76
 77
 78
 79
 80
 81
 82
 83
 84
 85
 86
 87
 88
 89
 90
 91
 92
 93
 94
 95
 96
 97
 98
 99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121

fn main() -> Result<()> {
  // Enable the multi-memory feature.
  let mut config = Config::new();
  config.wasm_multi_memory(true);

  let engine = Engine::new(&config)?;

  // Create our `store_fn` context and then compile a module and create an
  // instance from the compiled module all in one go.
  let mut store = Store::new(&engine, ());
  let module = Module::from_file(store.engine(), "examples/multimemory.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  let memory0 = instance
    .get_memory(&mut store, "memory0")
    .ok_or(anyhow::format_err!("failed to find `memory0` export"))?;
  let size0 = instance.get_typed_func::<(), i32, _>(&mut store, "size0")?;
  let load0 = instance.get_typed_func::<i32, i32, _>(&mut store, "load0")?;
  let store0 = instance.get_typed_func::<(i32, i32), (), _>(&mut store, "store0")?;

  let memory1 = instance
    .get_memory(&mut store, "memory1")
    .ok_or(anyhow::format_err!("failed to find `memory1` export"))?;
  let size1 = instance.get_typed_func::<(), i32, _>(&mut store, "size1")?;
  let load1 = instance.get_typed_func::<i32, i32, _>(&mut store, "load1")?;
  let store1 = instance.get_typed_func::<(i32, i32), (), _>(&mut store, "store1")?;

  println!("Checking memory...");
  assert_eq!(memory0.size(&store), 2);
  assert_eq!(memory0.data_size(&store), 0x20000);
  assert_eq!(memory0.data_mut(&mut store)[0], 0);
  assert_eq!(memory0.data_mut(&mut store)[0x1000], 1);
  assert_eq!(memory0.data_mut(&mut store)[0x1001], 2);
  assert_eq!(memory0.data_mut(&mut store)[0x1002], 3);
  assert_eq!(memory0.data_mut(&mut store)[0x1003], 4);

  assert_eq!(size0.call(&mut store, ())?, 2);
  assert_eq!(load0.call(&mut store, 0)?, 0);
  assert_eq!(load0.call(&mut store, 0x1000)?, 1);
  assert_eq!(load0.call(&mut store, 0x1001)?, 2);
  assert_eq!(load0.call(&mut store, 0x1002)?, 3);
  assert_eq!(load0.call(&mut store, 0x1003)?, 4);
  assert_eq!(load0.call(&mut store, 0x1ffff)?, 0);
  assert!(load0.call(&mut store, 0x20000).is_err()); // out of bounds trap

  assert_eq!(memory1.size(&store), 2);
  assert_eq!(memory1.data_size(&store), 0x20000);
  assert_eq!(memory1.data_mut(&mut store)[0], 0);
  assert_eq!(memory1.data_mut(&mut store)[0x1000], 4);
  assert_eq!(memory1.data_mut(&mut store)[0x1001], 3);
  assert_eq!(memory1.data_mut(&mut store)[0x1002], 2);
  assert_eq!(memory1.data_mut(&mut store)[0x1003], 1);

  assert_eq!(size1.call(&mut store, ())?, 2);
  assert_eq!(load1.call(&mut store, 0)?, 0);
  assert_eq!(load1.call(&mut store, 0x1000)?, 4);
  assert_eq!(load1.call(&mut store, 0x1001)?, 3);
  assert_eq!(load1.call(&mut store, 0x1002)?, 2);
  assert_eq!(load1.call(&mut store, 0x1003)?, 1);
  assert_eq!(load1.call(&mut store, 0x1ffff)?, 0);
  assert!(load0.call(&mut store, 0x20000).is_err()); // out of bounds trap

  println!("Mutating memory...");
  memory0.data_mut(&mut store)[0x1003] = 5;

  store0.call(&mut store, (0x1002, 6))?;
  assert!(store0.call(&mut store, (0x20000, 0)).is_err()); // out of bounds trap

  assert_eq!(memory0.data(&store)[0x1002], 6);
  assert_eq!(memory0.data(&store)[0x1003], 5);
  assert_eq!(load0.call(&mut store, 0x1002)?, 6);
  assert_eq!(load0.call(&mut store, 0x1003)?, 5);

  memory1.data_mut(&mut store)[0x1003] = 7;

  store1.call(&mut store, (0x1002, 8))?;
  assert!(store1.call(&mut store, (0x20000, 0)).is_err()); // out of bounds trap

  assert_eq!(memory1.data(&store)[0x1002], 8);
  assert_eq!(memory1.data(&store)[0x1003], 7);
  assert_eq!(load1.call(&mut store, 0x1002)?, 8);
  assert_eq!(load1.call(&mut store, 0x1003)?, 7);

  println!("Growing memory...");
  memory0.grow(&mut store, 1)?;
  assert_eq!(memory0.size(&store), 3);
  assert_eq!(memory0.data_size(&store), 0x30000);

  assert_eq!(load0.call(&mut store, 0x20000)?, 0);
  store0.call(&mut store, (0x20000, 0))?;
  assert!(load0.call(&mut store, 0x30000).is_err());
  assert!(store0.call(&mut store, (0x30000, 0)).is_err());

  assert!(memory0.grow(&mut store, 1).is_err());
  assert!(memory0.grow(&mut store, 0).is_ok());

  memory1.grow(&mut store, 2)?;
  assert_eq!(memory1.size(&store), 4);
  assert_eq!(memory1.data_size(&store), 0x40000);

  assert_eq!(load1.call(&mut store, 0x30000)?, 0);
  store1.call(&mut store, (0x30000, 0))?;
  assert!(load1.call(&mut store, 0x40000).is_err());
  assert!(store1.call(&mut store, (0x40000, 0)).is_err());

  assert!(memory1.grow(&mut store, 1).is_err());
  assert!(memory1.grow(&mut store, 0).is_ok());

  Ok(())
}

pub fn interrupt_handle(&self) -> Result<InterruptHandle, Error>

Creates an InterruptHandle which can be used to interrupt the execution of instances within this Store.

An InterruptHandle handle is a mechanism of ensuring that guest code doesn’t execute for too long. For example it’s used to prevent wasm programs for executing infinitely in infinite loops or recursive call chains.

The InterruptHandle type is sendable to other threads so you can interact with it even while the thread with this Store is executing wasm code.

There’s one method on an interrupt handle: InterruptHandle::interrupt. This method is used to generate an interrupt and cause wasm code to exit “soon”.

When are interrupts delivered?

The term “interrupt” here refers to one of two different behaviors that are interrupted in wasm:

The head of every loop in wasm has a check to see if it’s interrupted.

The prologue of every function has a check to see if it’s interrupted.

This interrupt mechanism makes no attempt to signal interrupts to native code. For example if a host function is blocked, then sending an interrupt will not interrupt that operation.

Interrupts are consumed as soon as possible when wasm itself starts executing. This means that if you interrupt wasm code then it basically guarantees that the next time wasm is executing on the target thread it will return quickly (either normally if it were already in the process of returning or with a trap from the interrupt). Once an interrupt trap is generated then an interrupt is consumed, and further execution will not be interrupted (unless another interrupt is set).

When implementing interrupts you’ll want to ensure that the delivery of interrupts into wasm code is also handled in your host imports and functionality. Host functions need to either execute for bounded amounts of time or you’ll need to arrange for them to be interrupted as well.

Return Value

This function returns a Result since interrupts are not always enabled. Interrupts are enabled via the Config::interruptable method, and if this store’s Config hasn’t been configured to enable interrupts then an error is returned.

Examples

// Enable interruptable code via `Config` and then create an interrupt
// handle which we'll use later to interrupt running code.
let engine = Engine::new(Config::new().interruptable(true))?;
let mut store = Store::new(&engine, ());
let interrupt_handle = store.interrupt_handle()?;

// Compile and instantiate a small example with an infinite loop.
let module = Module::new(&engine, r#"
    (func (export "run") (loop br 0))
"#)?;
let instance = Instance::new(&mut store, &module, &[])?;
let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

// Spin up a thread to send us an interrupt in a second
std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    interrupt_handle.interrupt();
});

let trap = run.call(&mut store, ()).unwrap_err();
assert!(trap.to_string().contains("wasm trap: interrupt"));

Examples found in repository

examples/interrupt.rs (line 14)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

fn main() -> Result<()> {
  // Enable interruptable code via `Config` and then create an interrupt
  // handle which we'll use later to interrupt running code.
  let engine = Engine::new(Config::new().interruptable(true))?;
  let mut store = Store::new(&engine, ());
  let interrupt_handle = store.interrupt_handle()?;

  // Compile and instantiate a small example with an infinite loop.
  let module = Module::from_file(&engine, "examples/interrupt.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // Spin up a thread to send us an interrupt in a second
  std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    println!("Interrupting!");
    interrupt_handle.interrupt();
  });

  println!("Entering infinite loop ...");
  let trap = run.call(&mut store, ()).unwrap_err();

  println!("trap received...");
  assert!(trap.to_string().contains("wasm trap: interrupt"));

  Ok(())
}

pub fn gc(&mut self)

Perform garbage collection of ExternRefs.

Note that it is not required to actively call this function. GC will automatically happen when internal buffers fill up. This is provided if fine-grained control over the GC is desired.

Examples found in repository

examples/externref.rs (line 53)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57

fn main() -> Result<()> {
  println!("Initializing...");
  let mut config = Config::new();
  config.wasm_reference_types(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());

  println!("Compiling module...");
  let module = Module::from_file(&engine, "examples/externref.wat")?;

  println!("Instantiating module...");
  let instance = Instance::new(&mut store, &module, &[])?;

  println!("Creating new `externref`...");
  let externref = ExternRef::new("Hello, World!");
  assert!(externref.data().is::<&'static str>());
  assert_eq!(
    *externref.data().downcast_ref::<&'static str>().unwrap(),
    "Hello, World!"
  );

  println!("Touching `externref` table...");
  let table = instance.get_table(&mut store, "table").unwrap();
  table.set(&mut store, 3, Some(externref.clone()).into())?;
  let elem = table
    .get(&mut store, 3)
    .unwrap() // assert in bounds
    .unwrap_externref() // assert it's an externref table
    .unwrap(); // assert the externref isn't null
  assert!(elem.ptr_eq(&externref));

  println!("Touching `externref` global...");
  let global = instance.get_global(&mut store, "global").unwrap();
  global.set(&mut store, Some(externref.clone()).into())?;
  let global_val = global.get(&mut store).unwrap_externref().unwrap();
  assert!(global_val.ptr_eq(&externref));

  println!("Calling `externref` func...");
  let func =
    instance.get_typed_func::<Option<ExternRef>, Option<ExternRef>, _>(&mut store, "func")?;
  let ret = func.call(&mut store, Some(externref.clone()))?;
  assert!(ret.is_some());
  assert!(ret.unwrap().ptr_eq(&externref));

  println!("GCing within the store...");
  store.gc();

  println!("Done.");
  Ok(())
}

pub fn fuel_consumed(&self) -> Option<u64>

Returns the amount of fuel consumed by this store’s execution so far.

If fuel consumption is not enabled via Config::consume_fuel then this function will return None. Also note that fuel, if enabled, must be originally configured via Store::add_fuel.

Examples found in repository

examples/fuel.rs (line 20)

≺ ≻↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

fn main() -> Result<()> {
  let mut config = Config::new();
  config.consume_fuel(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());
  store.add_fuel(10_000)?;
  let module = Module::from_file(store.engine(), "examples/fuel.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fibonacci` export with higher and higher numbers until we exhaust our fuel.
  let fibonacci = instance.get_typed_func::<i32, i32, _>(&mut store, "fibonacci")?;
  for n in 1.. {
    let fuel_before = store.fuel_consumed().unwrap();
    let output = match fibonacci.call(&mut store, n) {
      Ok(v) => v,
      Err(_) => {
        println!("Exhausted fuel computing fib({})", n);
        break;
      }
    };
    let fuel_consumed = store.fuel_consumed().unwrap() - fuel_before;
    println!("fib({}) = {} [consumed {} fuel]", n, output, fuel_consumed);
    store.add_fuel(fuel_consumed)?;
  }
  Ok(())
}

pub fn add_fuel(&mut self, fuel: u64) -> Result<(), Error>

Adds fuel to this Store for wasm to consume while executing.

For this method to work fuel consumption must be enabled via Config::consume_fuel. By default a Store starts with 0 fuel for wasm to execute with (meaning it will immediately trap). This function must be called for the store to have some fuel to allow WebAssembly to execute.

Most WebAssembly instructions consume 1 unit of fuel. Some instructions, such as nop, drop, block, and loop, consume 0 units, as any execution cost associated with them involves other instructions which do consume fuel.

Note that at this time when fuel is entirely consumed it will cause wasm to trap. More usages of fuel are planned for the future.

Panics

This function will panic if the store’s Config did not have fuel consumption enabled.

Examples found in repository

examples/fuel.rs (line 13)

≺ ≻↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

fn main() -> Result<()> {
  let mut config = Config::new();
  config.consume_fuel(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());
  store.add_fuel(10_000)?;
  let module = Module::from_file(store.engine(), "examples/fuel.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fibonacci` export with higher and higher numbers until we exhaust our fuel.
  let fibonacci = instance.get_typed_func::<i32, i32, _>(&mut store, "fibonacci")?;
  for n in 1.. {
    let fuel_before = store.fuel_consumed().unwrap();
    let output = match fibonacci.call(&mut store, n) {
      Ok(v) => v,
      Err(_) => {
        println!("Exhausted fuel computing fib({})", n);
        break;
      }
    };
    let fuel_consumed = store.fuel_consumed().unwrap() - fuel_before;
    println!("fib({}) = {} [consumed {} fuel]", n, output, fuel_consumed);
    store.add_fuel(fuel_consumed)?;
  }
  Ok(())
}

pub fn consume_fuel(&mut self, fuel: u64) -> Result<u64, Error>

Synthetically consumes fuel from this Store.

For this method to work fuel consumption must be enabled via Config::consume_fuel.

WebAssembly execution will automatically consume fuel but if so desired the embedder can also consume fuel manually to account for relative costs of host functions, for example.

This function will attempt to consume fuel units of fuel from within this store. If the remaining amount of fuel allows this then Ok(N) is returned where N is the amount of remaining fuel. Otherwise an error is returned and no fuel is consumed.

Errors

This function will return an either either if fuel consumption via Config is disabled or if fuel exceeds the amount of remaining fuel within this store.

pub fn out_of_fuel_trap(&mut self)

Configures a Store to generate a Trap whenever it runs out of fuel.

When a Store is configured to consume fuel with Config::consume_fuel this method will configure what happens when fuel runs out. Specifically a WebAssembly trap will be raised and the current execution of WebAssembly will be aborted.

This is the default behavior for running out of fuel.

pub fn out_of_fuel_async_yield(
&mut self,
injection_count: u64,
fuel_to_inject: u64
)

Configures a Store to yield execution of async WebAssembly code periodically.

When a Store is configured to consume fuel with Config::consume_fuel this method will configure what happens when fuel runs out. Specifically executing WebAssembly will be suspended and control will be yielded back to the caller. This is only suitable with use of a store associated with an async config because only then are futures used and yields are possible.

The purpose of this behavior is to ensure that futures which represent execution of WebAssembly do not execute too long inside their Future::poll method. This allows for some form of cooperative multitasking where WebAssembly will voluntarily yield control periodically (based on fuel consumption) back to the running thread.

Note that futures returned by this crate will automatically flag themselves to get re-polled if a yield happens. This means that WebAssembly will continue to execute, just after giving the host an opportunity to do something else.

The fuel_to_inject parameter indicates how much fuel should be automatically re-injected after fuel runs out. This is how much fuel will be consumed between yields of an async future.

The injection_count parameter indicates how many times this fuel will be injected. Multiplying the two parameters is the total amount of fuel this store is allowed before wasm traps.

Panics

This method will panic if it is not called on a store associated with an async config.

Examples found in repository

examples/tokio/main.rs (line 99)

↕

88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114

async fn run_wasm(inputs: Inputs) -> Result<(), Error> {
  let wasi = WasiCtxBuilder::new()
    // Let wasi print to this process's stdout.
    .inherit_stdout()
    // Set an environment variable so the wasm knows its name.
    .env("NAME", &inputs.name)?
    .build();
  let mut store = Store::new(&inputs.env.engine, wasi);

  // WebAssembly execution will be paused for an async yield every time it
  // consumes 10000 fuel. Fuel will be refilled u64::MAX times.
  store.out_of_fuel_async_yield(u64::MAX, 10000);

  // Instantiate into our own unique store using the shared linker, afterwards
  // acquiring the `_start` function for the module and executing it.
  let instance = inputs
    .env
    .linker
    .instantiate_async(&mut store, &inputs.env.module)
    .await?;
  instance
    .get_typed_func::<(), (), _>(&mut store, "_start")?
    .call_async(&mut store, ())
    .await?;

  Ok(())
}