Struct wiggle::wasmtime_crate::TypedFunc

#[repr(transparent)]
pub struct TypedFunc<Params, Results> { /* fields omitted */ }

A statically typed WebAssembly function.

Values of this type represent statically type-checked WebAssembly functions. The function within a TypedFunc is statically known to have Params as its parameters and Results as its results.

This structure is created via Func::typed or TypedFunc::new_unchecked. For more documentation about this see those methods.

Implementations

impl<Params, Results> TypedFunc<Params, Results> where
Params: WasmParams,
Results: WasmResults,

pub unsafe fn new_unchecked(func: Func) -> TypedFunc<Params, Results>

An unchecked version of Func::typed which does not perform a typecheck and simply assumes that the type declared here matches the type of this function.

The semantics of this function are the same as Func::typed except that no error is returned because no typechecking is done.

Unsafety

This function only safe to call if typed would otherwise return Ok for the same Params and Results specified. If typed would return an error then the returned TypedFunc is memory unsafe to invoke.

pub fn func(&self) -> &Func

Returns the underlying Func that this is wrapping, losing the static type information in the process.

pub fn call(
&self,
store: impl AsContextMut,
params: Params
) -> Result<Results, Trap>

Invokes this WebAssembly function with the specified parameters.

Returns either the results of the call, or a Trap if one happened.

For more information, see the Func::typed and Func::call documentation.

Panics

This function will panic if it is called when the underlying Func is connected to an asynchronous store.

Examples found in repository

examples/gcd.rs (line 20)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22

fn main() -> Result<()> {
  // Load our WebAssembly (parsed WAT in our case), and then load it into a
  // `Module` which is attached to a `Store` cache. After we've got that we
  // can instantiate it.
  let mut store = Store::<()>::default();
  let module = Module::from_file(store.engine(), "examples/gcd.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `gcd` export
  let gcd = instance.get_typed_func::<(i32, i32), i32, _>(&mut store, "gcd")?;

  println!("gcd(6, 27) = {}", gcd.call(&mut store, (6, 27))?);
  Ok(())
}

examples/fib-debug/main.rs (line 25)

↕

14
15
16
17
18
19
20
21
22
23
24
25
26
27

fn main() -> Result<()> {
  // Load our previously compiled wasm file (built previously with Cargo) and
  // also ensure that we generate debuginfo so this executable can be
  // debugged in GDB.
  let engine = Engine::new(Config::new().debug_info(true))?;
  let mut store = Store::new(&engine, ());
  let module = Module::from_file(&engine, "target/wasm32-unknown-unknown/debug/fib.wasm")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fib` export
  let fib = instance.get_typed_func::<i32, i32, _>(&mut store, "fib")?;
  println!("fib(6) = {}", fib.call(&mut store, 6)?);
  Ok(())
}

examples/threads.rs (line 59)

≺ ≻↕

49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70

fn run(engine: &Engine, module: &Module, linker: &Linker<()>) -> Result<()> {
  // Each sub-thread we have starting out by instantiating the `module`
  // provided into a fresh `Store`.
  println!("Instantiating module...");
  let mut store = Store::new(&engine, ());
  let instance = linker.instantiate(&mut store, module)?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  println!("Executing...");
  for _ in 0..N_REPS {
    run.call(&mut store, ())?;
    thread::sleep(time::Duration::from_millis(100));
  }

  // Also note that that a `Store` can also move between threads:
  println!("> Moving {:?} to a new thread", thread::current().id());
  let child = thread::spawn(move || run.call(&mut store, ()));

  child.join().unwrap()?;

  Ok(())
}

examples/wasi/main.rs (line 31)

↕

10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34

fn main() -> Result<()> {
  // Define the WASI functions globally on the `Config`.
  let engine = Engine::default();
  let mut linker = Linker::new(&engine);
  wasmtime_wasi::add_to_linker(&mut linker, |s| s)?;

  // Create a WASI context and put it in a Store; all instances in the store
  // share this context. `WasiCtxBuilder` provides a number of ways to
  // configure what the target program will have access to.
  let wasi = WasiCtxBuilder::new()
    .inherit_stdio()
    .inherit_args()?
    .build();
  let mut store = Store::new(&engine, wasi);

  // Instantiate our module with the imports we've created, and run it.
  let module = Module::from_file(&engine, "target/wasm32-wasi/debug/wasi.wasm")?;
  linker.module(&mut store, "", &module)?;
  linker
    .get_default(&mut store, "")?
    .typed::<(), (), _>(&store)?
    .call(&mut store, ())?;

  Ok(())
}

examples/fuel.rs (line 21)

↕

 8
 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33

fn main() -> Result<()> {
  let mut config = Config::new();
  config.consume_fuel(true);
  let engine = Engine::new(&config)?;
  let mut store = Store::new(&engine, ());
  store.add_fuel(10_000)?;
  let module = Module::from_file(store.engine(), "examples/fuel.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;

  // Invoke `fibonacci` export with higher and higher numbers until we exhaust our fuel.
  let fibonacci = instance.get_typed_func::<i32, i32, _>(&mut store, "fibonacci")?;
  for n in 1.. {
    let fuel_before = store.fuel_consumed().unwrap();
    let output = match fibonacci.call(&mut store, n) {
      Ok(v) => v,
      Err(_) => {
        println!("Exhausted fuel computing fib({})", n);
        break;
      }
    };
    let fuel_consumed = store.fuel_consumed().unwrap() - fuel_before;
    println!("fib({}) = {} [consumed {} fuel]", n, output, fuel_consumed);
    store.add_fuel(fuel_consumed)?;
  }
  Ok(())
}

examples/interrupt.rs (line 29)

↕

 9
10
11
12
13
14
15
16
17
18
19
20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35

fn main() -> Result<()> {
  // Enable interruptable code via `Config` and then create an interrupt
  // handle which we'll use later to interrupt running code.
  let engine = Engine::new(Config::new().interruptable(true))?;
  let mut store = Store::new(&engine, ());
  let interrupt_handle = store.interrupt_handle()?;

  // Compile and instantiate a small example with an infinite loop.
  let module = Module::from_file(&engine, "examples/interrupt.wat")?;
  let instance = Instance::new(&mut store, &module, &[])?;
  let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;

  // Spin up a thread to send us an interrupt in a second
  std::thread::spawn(move || {
    std::thread::sleep(std::time::Duration::from_secs(1));
    println!("Interrupting!");
    interrupt_handle.interrupt();
  });

  println!("Entering infinite loop ...");
  let trap = run.call(&mut store, ()).unwrap_err();

  println!("trap received...");
  assert!(trap.to_string().contains("wasm trap: interrupt"));

  Ok(())
}

Additional examples can be found in:

examples/linking.rs

examples/multi.rs

examples/serialize.rs

examples/externref.rs

examples/hello.rs

examples/memory.rs

examples/multimemory.rs

pub async fn call_async<T>(
&'_ self,
store: impl AsContextMut<Data = T>,
params: Params
) -> Result<Results, Trap> where
T: Send,

Invokes this WebAssembly function with the specified parameters.

Returns either the results of the call, or a Trap if one happened.

For more information, see the Func::typed and Func::call_async documentation.

Panics

This function will panic if it is called when the underlying Func is connected to a synchronous store.

Examples found in repository

examples/tokio/main.rs (line 110)

↕

88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114

async fn run_wasm(inputs: Inputs) -> Result<(), Error> {
  let wasi = WasiCtxBuilder::new()
    // Let wasi print to this process's stdout.
    .inherit_stdout()
    // Set an environment variable so the wasm knows its name.
    .env("NAME", &inputs.name)?
    .build();
  let mut store = Store::new(&inputs.env.engine, wasi);

  // WebAssembly execution will be paused for an async yield every time it
  // consumes 10000 fuel. Fuel will be refilled u64::MAX times.
  store.out_of_fuel_async_yield(u64::MAX, 10000);

  // Instantiate into our own unique store using the shared linker, afterwards
  // acquiring the `_start` function for the module and executing it.
  let instance = inputs
    .env
    .linker
    .instantiate_async(&mut store, &inputs.env.module)
    .await?;
  instance
    .get_typed_func::<(), (), _>(&mut store, "_start")?
    .call_async(&mut store, ())
    .await?;

  Ok(())
}