Struct wiggle::wasmtime_crate::V128

pub struct V128(/* private fields */);

Representation of a 128-bit vector type, v128, for WebAssembly.

This type corresponds to the v128 type in WebAssembly and can be used with the TypedFunc API for example. This is additionally the payload of Val::V128.

Platform specifics

This type can currently only be used on x86_64 and AArch64 with the TypedFunc API. Rust does not have stable support on other platforms for this type so invoking functions with v128 parameters requires the Func::call API (or perhaps Func::call_unchecked.

Implementations

impl V128

pub fn as_u128(&self) -> u128

Returns the representation of this v128 as a 128-bit integer in Rust.