Trait wiggle::wasmtime_crate::RootedGcRef

pub trait RootedGcRef<T>: RootedGcRefImpl<T> + Deref<Target = T>
where
    T: GcRef,
{ }

A trait implemented for GC references that are guaranteed to be rooted:

Rooted<T>

ManuallyRooted<T>

You can use this to abstract over the different kinds of rooted GC references. Note that Deref<Target = T> is a supertrait for RootedGcRef<T>, so all rooted GC references deref to their underlying T, allowing you to call its methods.

This is a sealed trait, and may not be implemented for any types outside of the wasmtime crate.