Trait wiggle::wasmtime_crate::WasmRet

pub unsafe trait WasmRet { }

A trait implemented for types which can be returned from closures passed to Func::wrap and friends.

This trait should not be implemented by user types. This trait may change at any time internally. The types which implement this trait, however, are stable over time.

For more information see Func::wrap