Crate wiggle

Re-exports

pub use witx;

Macros

from_witx
This macro expands to a set of pub Rust modules:

Structs

GuestBorrows

GuestPtr
A guest pointer into host memory.

Region
Represents a contiguous region in memory.

Enums

GuestError

Traits

GuestErrorType
A trait for types which are used to report errors. Each type used in the first result position of an interface function is used, by convention, to indicate whether the function was successful and subsequent results are valid, or whether an error occured. This trait allows wiggle to return the correct value when the interface function's idiomatic Rust method returns Ok().

GuestMemory
A trait which abstracts how to get at the region of host memory taht contains guest memory.

GuestType
A trait for types that are intended to be pointees in GuestPtr<T>.

GuestTypeTransparent
A trait for GuestTypes that have the same representation in guest memory as in Rust. These types can be used with the GuestPtr::as_raw method to view as a slice.

Pointee
Types that can be pointed to by GuestPtr<T>.