Module wiggle::wasmtime_crate

Modules

component

Embedding API for the Component Model

unix

Unix-specific extension for the wasmtime crate.

Structs

Caller

A structure representing the caller’s context when creating a function via Func::wrap.

CodeBuilder

Builder-style structure used to create a Module or pre-compile a module to a serialized list of bytes.

CodeMemory

Management of executable memory within a MmapVec

CompiledModule

A compiled wasm module, ready to be instantiated.

Config

Global configuration options used to create an Engine and customize its behavior.

Engine

An Engine which is a global context for compilation and management of wasm modules.

EngineWeak

A weak reference to an Engine.

Error

The Error type, a wrapper around a dynamic error type.

Export

An exported WebAssembly value.

ExportType

A descriptor for an exported WebAssembly value.

ExternRef

An opaque, GC-managed reference to some host data that can be passed to WebAssembly.

FrameInfo

Description of a frame in a backtrace for a WasmBacktrace.

FrameSymbol

Debug information for a symbol that is attached to a FrameInfo.

Func

A WebAssembly function which can be called.

FuncType

The type of a WebAssembly function.

Global

A WebAssembly global value which can be read and written to.

GlobalType

A WebAssembly global descriptor.

GuestProfiler

Collects basic profiling data for a single WebAssembly guest.

ImportType

A descriptor for an imported value into a wasm module.

Instance

An instantiated WebAssembly module.

InstancePre

An instance, pre-instantiation, that is ready to be instantiated.

Linker

Structure used to link wasm modules/instances together.

ManuallyRooted

A rooted reference to a garbage-collected T with arbitrary lifetime.

Memory

A WebAssembly linear memory.

MemoryAccessError

Error for out of bounds Memory access.

MemoryType

A descriptor for a WebAssembly memory type.

Module

A compiled WebAssembly module, ready to be instantiated.

ModuleExport

Describes the location of an export in a module.

NoFunc

A reference to the abstract nofunc heap value.

PoolingAllocationConfig

Configuration options used with InstanceAllocationStrategy::Pooling to change the behavior of the pooling instance allocator.

RefType

Opaque references to data in the Wasm heap or to host data.

ResourcesRequired

A summary of the amount of resources required to instantiate a particular Module or Component.

RootScope

Nested rooting scopes.

Rooted

A scoped, rooted reference to a garbage-collected T.

SharedMemory

A constructor for externally-created shared memory.

Store

A Store is a collection of WebAssembly instances and host-defined state.

StoreContext

A temporary handle to a &Store<T>.

StoreContextMut

A temporary handle to a &mut Store<T>.

StoreLimits

Provides limits for a Store.

StoreLimitsBuilder

Used to build StoreLimits.

Table

A WebAssembly table, or an array of values.

TableType

A descriptor for a table in a WebAssembly module.

TypedFunc

A statically typed WebAssembly function.

UnknownImportError

Error for an unresolvable import.

V128

Representation of a 128-bit vector type, v128, for WebAssembly.

WasmBacktrace

Representation of a backtrace of function frames in a WebAssembly module for where an error happened.

WasmCoreDump

Representation of a core dump of a WebAssembly module

Enums

CallHook

Passed to the argument of Store::call_hook to indicate a state transition in the WebAssembly VM.

Extern

An external item to a WebAssembly module, or a list of what can possibly be exported from a wasm module.

ExternType

A list of all possible types which can be externally referenced from a WebAssembly module.

HeapType

The heap types that can Wasm can have references to.

InstanceAllocationStrategy

Represents the module instance allocation strategy to use.

ModuleVersionStrategy

Configure the strategy used for versioning in serializing and deserializing crate::Module.

MpkEnabled

Describe the tri-state configuration of memory protection keys (MPK).

Mutability

Indicator of whether a global is mutable or not

OptLevel

Possible optimization levels for the Cranelift codegen backend.

Precompiled

Return value from the Engine::detect_precompiled API.

ProfilingStrategy

Select which profiling technique to support.

Ref

A reference.

Strategy

Possible Compilation strategies for a wasm module.

Trap

UpdateDeadline

What to do after returning from a callback when the engine epoch reaches the deadline for a Store during execution of a function using that store.

Val

Possible runtime values that a WebAssembly module can either consume or produce.

ValType

A list of all possible value types in WebAssembly.

WaitResult

Result of Memory::atomic_wait32 and Memory::atomic_wait64

WasmBacktraceDetails

Select how wasm backtrace detailed information is handled.

Constants

DEFAULT_INSTANCE_LIMIT

Value returned by ResourceLimiter::instances default method

DEFAULT_MEMORY_LIMIT

Value returned by ResourceLimiter::memories default method

DEFAULT_TABLE_LIMIT

Value returned by ResourceLimiter::tables default method

Traits

AsContext

A trait used to get shared access to a Store in Wasmtime.

AsContextMut

A trait used to get exclusive mutable access to a Store in Wasmtime.

CallHookHandler

An object that can take callbacks when the runtime enters or exits hostcalls.

GcRef

A common trait implemented by all garbage-collected reference types.

IntoFunc

Internal trait implemented for all arguments that can be passed to Func::wrap and Linker::func_wrap.

LinearMemory

A linear memory. This trait provides an interface for raw memory buffers which are used by wasmtime, e.g. inside [‘Memory’]. Such buffers are in principle not thread safe. By implementing this trait together with MemoryCreator, one can supply wasmtime with custom allocated host managed memory.

MemoryCreator

A memory creator. Can be used to provide a memory creator to wasmtime which supplies host managed memory.

ResourceLimiter

Used by hosts to limit resource consumption of instances.

ResourceLimiterAsync

Used by hosts to limit resource consumption of instances, blocking asynchronously if necessary.

RootedGcRef

A trait implemented for GC references that are guaranteed to be rooted:

StackCreator

A stack creator. Can be used to provide a stack creator to wasmtime which supplies stacks for async support.

StackMemory

A stack memory. This trait provides an interface for raw memory buffers which are used by wasmtime inside of stacks which wasmtime executes WebAssembly in for async support. By implementing this trait together with StackCreator, one can supply wasmtime with custom allocated host managed stacks.

WasmParams

A trait used for Func::typed and with TypedFunc to represent the set of parameters for wasm functions.

WasmResults

A trait used for Func::typed and with TypedFunc to represent the set of results for wasm functions.

WasmRet

A trait implemented for types which can be returned from closures passed to Func::wrap and friends.

WasmTy

A trait implemented for types which can be arguments and results for closures passed to Func::wrap as well as parameters to Func::typed.

Type Aliases

Result

Result<T, Error>

Unions

ValRaw

A “raw” and unsafe representation of a WebAssembly value.