Module wiggle::wasmtime_crate::component

In-progress implementation of the WebAssembly component model

This module is a work-in-progress and currently represents an incomplete and probably buggy implementation of the component model.

Modules

types

This module defines the Type type, representing the dynamic form of a component interface type.

Macros

bindgen

flags

Structs

Component

A compiled WebAssembly Component.

ExportInstance

Description of the exports of a single instance.

Exports

Description of the exports of an Instance.

Func

A WebAssembly component function.

Instance

An instantiated component.

InstancePre

A “pre-instantiated” Instance which has all of its arguments already supplied and is ready to instantiate.

Linker

A type used to instantiate Components.

LinkerInstance

Structure representing an “instance” being defined within a linker.

TypedFunc

A statically-typed version of Func which takes Params as input and returns Return.

WasmList

Representation of a list of values that are owned by a WebAssembly instance.

WasmStr

Representation of a string located in linear memory in a WebAssembly instance.

Enums

Type

Represents a component model interface type

Val

Represents possible runtime values which a component function can either consume or produce

Traits

ComponentNamedList

A trait representing a static list of named types that can be passed to or returned from a TypedFunc.

ComponentType

A trait representing types which can be passed to and read from components with the canonical ABI.

Lift

Host types which can be created from the canonical ABI.

Lower

Host types which can be passed to WebAssembly components.

Derive Macros

ComponentType

Lift

Lower