Trait wiggle::wasmtime_crate::component::ComponentType

pub unsafe trait ComponentType { }

A trait representing types which can be passed to and read from components with the canonical ABI.

This trait is implemented for Rust types which can be communicated to components. This is implemented for Rust types which correspond to interface types in the component model of WebAssembly. The Func::typed and TypedFunc Rust items are the main consumers of this trait.

For more information on this trait see the examples in Func::typed.

The contents of this trait are hidden as it’s intended to be an implementation detail of Wasmtime. The contents of this trait are not covered by Wasmtime’s stability guarantees.